+++ src/core/component/render/CHANGELOG.md
@@ -9,6 +9,25 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v4.0.0-beta.151 (2024-11-06)
+
+#### :bug: Bug Fix
+
+* Fixed the `resolveAttrs` function: property getters are no longer removed from props, the `v-attrs` directive now resolves with the correct method in SSR
+* Calls `resolveAttrs` to resolve directives for components rendered with `ssrRenderComponent`
+
+## v4.0.0-beta.149 (2024-10-31)
+
+#### :rocket: New Feature
+
+* Added a wrapper for `withModifiers` with support for the `safe` modifier
+
+## v4.0.0-beta.148 (2024-10-28)
+
+#### :house: Internal
+
+* Create a `normalizeComponentForceUpdateProps` for normalizing the props with `forceUpdate = false`
+
 ## v4.0.0-beta.140 (2024-10-03)
 
 #### :bug: Bug Fix